requirements.txt
fastapi
uvicorn
pillow
transformers
torch
sentence-transformers
pymupdf
chromadb
langchain
openai
python-dotenv
jupyter
opencv-python
matplotlib
seaborn